move stories to pages folder, remove files & rename
